refactor(sql): move sql interface from repository
